==> Pre-Processing/Step4_automatic_feature_extraction.ipynb <==
# Feature Extraction and Selection

import six, numpy as np
import pandas as pd
import csv


import radiomics
from radiomics import featureextractor  # This module is used for interaction with pyradiomics
from collections import OrderedDict

import matplotlib.pyplot as plt
import seaborn as sb
import seaborn as sns

## Feature Extraction

Using an extractor to set up with the correct parameters and saving the results in a csv file

a = pd.read_csv('tumours_info_nii.csv')
dataPath = '../../data/nii/'

def f_extraction():
    # create the extractor just once outside of the loop to maybe safe processing time
    paramPath = '../../data/Params/Params2.yaml'
    extractor = featureextractor.RadiomicsFeatureExtractor(paramPath)
    
    # list to store the results
    results = []
    for ind in a.index:
        print("Processing case ", ind)
        scanPath = dataPath + a['Path'][ind]+ '/'+ a['Scan'][ind]
        maskPath = dataPath + a['Path'][ind]+ '/'+ a['GTV Primary'][ind]
        patient = a['Patient'][ind]
        print("Scan: ", scanPath)
        print("Mask: ", maskPath)
        # identify the patient  
        feature_vector = OrderedDict([('Patient', patient)])
        # add the features
        feature_vector.update(extractor.execute(scanPath, maskPath))
        # append to results list
        results.append(feature_vector)
            
        mask2 = a['GTV Primary 2'][ind]
        if mask2 == '-':
            pass
        else:
            #scanPath = dataPath + a['Path'][ind]+ '/'+ a['Scan'][ind]
            maskPath2 = dataPath + a['Path'][ind]+ '/'+ a['GTV Primary 2'][ind]
            patient = a['Patient'][ind]
            #print("Scan: ", scanPath)
            print("Mask2: ", maskPath)
            # identify the patient  
            feature_vector = OrderedDict([('Patient', patient)])
            # add the features
            feature_vector.update(extractor.execute(scanPath, maskPath2))
            # append to results list
            results.append(feature_vector)
                

    print("Done processing. Writing to file...")
    # initialize the dataframe with the first result
    data_frame = pd.DataFrame.from_dict(results[0], orient="index")
    
    # iterate over the remaining results
    for idx, dict in enumerate(results):
        if idx == 0:
            continue
        df = pd.DataFrame.from_dict(dict, orient="index")
        data_frame = pd.concat([data_frame, df], axis=1)
        data_frame_T = data_frame.T
    data_frame_T.to_csv('All_Filters_Features.csv', index=False)
    print("Done!")
            
f_extraction()

#test: if ind <5 or 8<ind<22 :
#train: if ind>21 :
