==> Pre-Processing/Step5_Normalization.ipynb <==
# Normalization

## 1- Radiomic Features

import pandas as pd

#Passo1- Limpeza das features extraídas

def clean_radiomic():
    ExtractedFeatures = pd.read_csv("Features.csv")
    
    clean = ExtractedFeatures.drop(columns=['diagnostics_Versions_PyRadiomics',
                                            'diagnostics_Versions_Numpy', 'diagnostics_Versions_SimpleITK',
                                            'diagnostics_Versions_PyWavelet', 'diagnostics_Versions_Python',
                                            'diagnostics_Configuration_Settings',
                                            'diagnostics_Configuration_EnabledImageTypes',
                                            'diagnostics_Image-original_Hash',
                                            'diagnostics_Image-original_Dimensionality',
                                            'diagnostics_Image-original_Spacing',
                                            'diagnostics_Image-original_Size', 
                                            'diagnostics_Mask-original_Spacing',
                                            'diagnostics_Mask-original_Size', 
                                            'diagnostics_Mask-original_BoundingBox',
                                            'diagnostics_Mask-original_Hash', 
                                            'diagnostics_Mask-original_CenterOfMassIndex',
                                            'diagnostics_Mask-original_CenterOfMass'])
    
    clean.to_csv("Features_Clean.csv", index=False)
    
    
    
clean_radiomic()  

'''
#Passo2- Normalizar o ficheiro com as features extraidas

def normalize_radiomic():
    df = pd.read_csv('Clean_Features.csv')
    
    data = df.drop(columns= ['Patient'])
    
    data_normalized = data.apply(lambda x:( (x - x.min()) / (x.max()-x.min())))
    
    #data_normalized.to_csv('Normalized_All_Filters_Features.csv', index=False)
    
    return data_normalized
   
normalize_radiomic()    
'''

# apply the min-max scaling in Pandas using the .min() and .max() methods
#Assim por coluna já está correto!
def min_max_scaling(df):
    # read the dataframe
    df_norm = df.drop(columns= ['Patient'])
    # apply min-max scaling
    for column in df_norm.columns:
        df_norm[column] = (df_norm[column] - df_norm[column].min()) / (df_norm[column].max() - df_norm[column].min())
    
    #
    
    return df_norm

df = pd.read_csv('All_Filters_Features_Clean.csv')
df=min_max_scaling(df)
df.to_csv('All_Filters_Features_Clean_Normalized.csv', index=False)   


def clean_radiomic_all_filters_features_clean_normalized():
    data = pd.read_csv("All_Filters_Features_Clean_Normalized.csv")
    
    clean = data.dropna( axis = 1, how ='any')
    
    clean.to_csv("All_Filters_Features_Clean_Normalized.csv", index=False)
    
    print(clean.shape)
    
clean_radiomic_all_filters_features_clean_normalized() 

#Passo3- Limpeza nas variáveis clínicas  
#desta vez fiz este passo à mão 
'''
def clean_clinical_data():
    df = pd.read_csv('clinical_data_nii.csv')
    
    
    df['Time – treatment (days)'] = df['Time – diagnosis to end treatment (days)'] - df['Time – diagnosis to start treatment (days)']
      
    
    a= df.drop(columns= ['M-stage', 'Time – diagnosis to diagnosis (days)', 
                     'Time – diagnosis to DM (days)', 'Time – diagnosis to Death (days)',
                     'Time – diagnosis to LR (days)', 'Surgery','Time – diagnosis to PET (days)',
                     'Time – diagnosis to end treatment (days)'])
       
    a= a[['Patient', 'Sex', 'Age', 'Primary Site', 'T-stage', 'N-stage',
                  'TNM group stage', 'HPV status', 'Lymph Nodes', 'Time – diagnosis to CT sim (days)',
                  'Time – diagnosis to start treatment (days)',
                  'Time – treatment (days)', 'Therapy',
                  'Time – diagnosis to last follow-up(days)', 'Locoregional', 'Distant', 'Death']]
    
    b = a.drop([5,6,7,8])
     
    b.to_csv('Clean_Clinical_Data.csv', index=False)
    
    return b

clean_clinical_data()

'''

#Passo4- Categorizar as variáveis clínicas
#hot enconding (sklearn) OU get.dummies (pandas)

def categorize_clinical_data():
    df= pd.read_csv('Clinical_Data_nii.csv')
    
    cat_columns = ["Sex", "Primary Site", "T-Stage", "N-Stage", "TNM group Stage", "Therapy", "HPV status", "Surgery"]
    
    df_processed = pd.get_dummies(df, columns=cat_columns)
    
    
    df_processed = df_processed[['Patient', 'Sex_F',  'Sex_M', 'Age', 'Primary Site_Hypopharynx', 
                                 'Primary Site_Larynx', 'Primary Site_Nasopharynx', 'Primary Site_Oropharynx', 
                                 'Primary Site_unknown', 'T-Stage_T1', 'T-Stage_T2', 'T-Stage_T3', 'T-Stage_T4', 
                                 'T-Stage_T4a', 'T-Stage_T4b', 'T-Stage_Tx', 'N-Stage_N0', 'N-Stage_N1', 'N-Stage_N2', 'N-Stage_N2a',
                                 'N-Stage_N2b', 'N-Stage_N2c', 'N-Stage_N3', 'N-Stage_N3b',
                                 'TNM group Stage_Stage I', 'TNM group Stage_Stage II', 'TNM group Stage_Stage IIB',
                                 'TNM group Stage_Stage III', 'TNM group Stage_Stage IV', 'TNM group Stage_Stage IVA',
                                 'TNM group Stage_Stage IVB', 'TNM group Stage_unknown', 'HPV status_0', 'HPV status_1',
                                 'HPV status_unknown', 'Lymph Nodes', 'Surgery_0', 'Surgery_1', 'Surgery_unknown',
                                 'Time – diagnosis to CT sim (days)', 'Time – diagnosis to start treatment (days) ', 'Time – treatment (days)', 
                                 'Time – diagnosis to last follow-up(days)', 'Therapy_chemo radiation', 'Therapy_radiation',
                                 'Locoregional', 'Distant','Death']]
    
    
    columns= ['Age', 'Time – treatment (days)', 'Time – diagnosis to last follow-up(days)']
    for column in columns:
        df_processed[column] = (df_processed[column] - df_processed[column].min()) / (df_processed[column].max() - df_processed[column].min())
    

    df_processed.to_csv('Clinical_Data_nii_final.csv', index=False)

    

    
categorize_clinical_data()   

def negative_values():
    df= pd.read_csv('Clinical_Data_nii_final_final.csv')
    
    
    
    columns=['Time – diagnosis to CT sim (days)', 'Time – diagnosis to start treatment (days) ']
    for column in columns:
        if df['Time – diagnosis to CT sim (days)'].str.contains("unknown") == True:
             pass
        else:
            df[column] = (df[column] - df[column].min()) / (df[column].max() - df[column].min())
    return df
       
    
    #print(df['Time – diagnosis to CT sim (days)'].str.contains("unknown"))
    #df[columns].values == 'unknown':    
    #print((df['Time – diagnosis to CT sim (days)']=='unknown').any())
    #df.to_csv('Clinical_Data_nii_final_final1.csv', index=False)
    
negative_values()    

#Passo5- Juntar as 3 labels ao 1 ficheiro (Normalized_Features)

def Normalized_Features_Labels():
    df = pd.read_csv('All_Filters_Features_Clean_Normalized.csv')
    
    df1= pd.read_csv('Clinical_Data_nii.csv')
    
    result = pd.concat([df, df1], axis=1)
    
    final= result.drop(columns= ['Patient', 'Sex', 'Age', 'Primary Site', 'T-stage', 'N-stage',
                                 'TNM group stage', 'Time – diagnosis to CT sim (days)',
                                 'Time – diagnosis to start treatment (days) ',
                                 'Time – treatment (days)', 'Therapy',
                                 'Time – diagnosis to last follow-up(days)', 'HPV status', 'Lymph Nodes', 'Surgery'])
    
    final.to_csv('All_Filters_Features_Clean_Normalized_Labels.csv', index=False)
    
    
    return final

Normalized_Features_Labels()

import pandas as pd

tumour_info = pd.read_csv("tumours_info_nii.csv")
tumour_info.head()
