==> Pre-Processing/Step5.1_Train+Test.ipynb <==
import pandas as pd

# Data Preparation

Os modelos vão ser treinados separadamente, ou seja, 3 modelos diferentes: LR, DM e OS.

Os modelos LR e DM são independentes, isto é, LR não influencia DM e vice-versa. No entanto, OS depende dos anteriores.

Para os dados de treino vão-se usar os dados de HGJ e CHUS
Para os dados de teste vão-se usar os dados de HMR e CHUM

## Locoregional Recurrences (LR)

#Passo1- ler ficheiro csv
#Passo2- eliminar as colunas relativas a DM e OS
#Passo3- dividir dataset em treino e teste 
#Passo4- guardar em ficheiros csv  distintos

def LR():
    Feat_Labels = pd.read_csv("All_Filters_Features_Clean_Normalized_Labels.csv")
    
    clean_cols = Feat_Labels.drop(columns=['Distant', 'Death'])
    
    train_rows = clean_cols.loc[32:156]

    
    test_rows1 = clean_cols.loc[0:31]
    test_rows2 = clean_cols.loc[157:]
    
    test_rows3 = [test_rows1, test_rows2]
    
    
    test_rows = pd.concat(test_rows3)
    
    #print(len(train_rows))
    #print(len(test_rows1))
    #print(len(test_rows2))
    #print(len(test_rows))
    train_rows.to_csv("All_Filters_Features_Clean_Normalized_LR_train.csv", index=False)
    test_rows.to_csv("All_Filters_Features_Clean_Normalized_LR_test.csv", index=False)
    
LR()

## Distant Metastasis (DM)

#Passo1- ler ficheiro csv
#Passo2- eliminar as colunas relativas a LR e OS
#Passo3- dividir dataset em treino e teste 
#Passo4- guardar em ficheiros csv  distintos

def DM():
    Feat_Labels = pd.read_csv("All_Filters_Features_Clean_Normalized_Labels.csv")
    
    clean_cols = Feat_Labels.drop(columns=['Locoregional', 'Death'])
    
    train_rows = clean_cols.loc[32:156]

    
    test_rows1 = clean_cols.loc[0:31]
    test_rows2 = clean_cols.loc[157:]
    
    test_rows3 = [test_rows1, test_rows2]
    
    
    test_rows = pd.concat(test_rows3)
    
    print(len(train_rows))
    print(len(test_rows1))
    print(len(test_rows2))
    print(len(test_rows))
    train_rows.to_csv("All_Filters_Features_Clean_Normalized_DM_train.csv", index=False)
    test_rows.to_csv("All_Filters_Features_Clean_Normalized_DM_test.csv", index=False)
    
DM()

## Overall Survival (OS)

#Passo1- ler ficheiro csv
#Passo2- dividir dataset em treino e teste 
#Passo3- guardar em ficheiros csv  distintos

def OS():
    Feat_Labels = pd.read_csv("All_Filters_Features_Clean_Normalized_Labels.csv")
    
    
    train_rows = Feat_Labels.loc[32:156]

    
    test_rows1 = Feat_Labels.loc[0:31]
    test_rows2 = Feat_Labels.loc[157:]
    
    test_rows3 = [test_rows1, test_rows2]
    
    
    test_rows = pd.concat(test_rows3)
    
    print(len(train_rows))
    print(len(test_rows1))
    print(len(test_rows2))
    print(len(test_rows))
    train_rows.to_csv("All_Filters_Features_Clean_Normalized_OS_train.csv", index=False)
    test_rows.to_csv("All_Filters_Features_Clean_Normalized_OS_test.csv", index=False)
    
OS()

Caso não seja para usar a mesma divisão que Martin Vallières, a divisão em casos de treino e teste é feita aleatoriamente, estando estes no mesmo ficheiro:

def LR_1():
    Feat_Labels = pd.read_csv("All_Filters_Features_Clean_Normalized_Labels.csv")
    
    clean_cols = Feat_Labels.drop(columns=['Distant', 'Death'])
    
    clean_cols.to_csv("All_Filters_Features_Clean_Normalized_LR.csv", index=False)
    
LR_1()

# Clinical Data

Features = pd.read_csv('Features_Clean_Normalized.csv')
Clinical_Data= pd.read_csv('Clinical_Data_nii_final.csv')

#Passo1- ler ficheiro csv features + clinical data
#Passo2- eliminar colunas no clinical data: pacientes + 4tempos + DM+ OS
#Passo3- juntar 2 csv
#Passo4- dividir em teste + treino


def clinical_data_LR(Features,Clinical_Data):
        
    clean_data = Clinical_Data.drop(columns=['Patient','Time – diagnosis to CT sim (days)', 
                                             'Time – diagnosis to start treatment (days) ', 
                                             'Time – treatment (days)', 
                                             'Time – diagnosis to last follow-up(days)', 
                                             'Distant', 'Death'])
    

    
    data = Features.join(clean_data, how='inner')
    
    print(len(data))
    
    
    train_rows= data.loc[32:156]

    
    test_rows1 = data.loc[0:31]
    test_rows2 = data.loc[157:]
    
    test_rows3 = [test_rows1, test_rows2]
    
    
    test_rows = pd.concat(test_rows3)
    
    print(len(train_rows))
    print(len(test_rows1))
    print(len(test_rows2))
    print(len(test_rows))
    
    train_rows.to_csv("Features_Clean_Normalized_Clinical_Data_LR_train.csv", index=False)
    test_rows.to_csv("Features_Clean_Normalized_Clinical_Data_LR_test.csv", index=False)
    
clinical_data_LR(Features,Clinical_Data)

#Passo1- ler ficheiro csv features + clinical data
#Passo2- eliminar colunas no clinical data: pacientes + 4tempos + DM+ OS
#Passo3- juntar 2 csv
#Passo4- dividir em teste + treino


def clinical_data_DM(Features,Clinical_Data):
        
    clean_data = Clinical_Data.drop(columns=['Patient','Time – diagnosis to CT sim (days)', 
                                             'Time – diagnosis to start treatment (days) ', 
                                             'Time – treatment (days)', 
                                             'Time – diagnosis to last follow-up(days)', 
                                             'Locoregional', 'Death'])
    

    
    data = Features.join(clean_data, how='inner')
    
    print(len(data))
    
    
    train_rows= data.loc[32:156]

    
    test_rows1 = data.loc[0:31]
    test_rows2 = data.loc[157:]
    
    test_rows3 = [test_rows1, test_rows2]
    
    
    test_rows = pd.concat(test_rows3)
    
    print(len(train_rows))
    print(len(test_rows1))
    print(len(test_rows2))
    print(len(test_rows))
    
    train_rows.to_csv("Features_Clean_Normalized_Clinical_Data_DM_train.csv", index=False)
    test_rows.to_csv("Features_Clean_Normalized_Clinical_Data_DM_test.csv", index=False)
    
clinical_data_DM(Features,Clinical_Data)

#Passo1- ler ficheiro csv features + clinical data
#Passo2- eliminar colunas no clinical data: pacientes + 4tempos + DM+ OS
#Passo3- juntar 2 csv
#Passo4- dividir em teste + treino

def clinical_data_OS(Features,Clinical_Data):
        
    clean_data = Clinical_Data.drop(columns=['Patient','Time – diagnosis to CT sim (days)', 
                                             'Time – diagnosis to start treatment (days) ', 
                                             'Time – treatment (days)', 
                                             'Time – diagnosis to last follow-up(days)'])
    
    data = Features.join(clean_data, how='inner')
       
    train_rows= data.loc[32:156]
   
    test_rows1 = data.loc[0:31]
    test_rows2 = data.loc[157:]
    
    test_rows3 = [test_rows1, test_rows2]
       
    test_rows = pd.concat(test_rows3)
    
    print(len(train_rows))
    print(len(test_rows1))
    print(len(test_rows2))
    print(len(test_rows))
    
    train_rows.to_csv("Features_Clean_Normalized_Clinical_Data_OS_train.csv", index=False)
    test_rows.to_csv("Features_Clean_Normalized_Clinical_Data_OS_test.csv", index=False)
    
clinical_data_OS(Features,Clinical_Data)

# Clinical Data - with time columns

Features = pd.read_csv('Features_Clean_Normalized.csv')
Clinical_Data= pd.read_csv('Clinical_Data_time.csv')

#Passo1- ler ficheiro csv features + clinical data
#Passo2- eliminar colunas no clinical data: pacientes + 4tempos + DM+ OS
#Passo3- juntar 2 csv
#Passo4- dividir em teste + treino


def clinical_data_LR(Features,Clinical_Data):
        
    clean_data = Clinical_Data.drop(columns=['Patient','Time – diagnosis to CT sim (days)', 'Distant', 'Death'])
    

    
    data = Features.join(clean_data, how='inner')
    
    print(len(data))
    
    
    train_rows= data.loc[32:156]

    
    test_rows1 = data.loc[0:31]
    test_rows2 = data.loc[157:]
    
    test_rows3 = [test_rows1, test_rows2]
    
    
    test_rows = pd.concat(test_rows3)
    
    print(len(train_rows))
    print(len(test_rows1))
    print(len(test_rows2))
    print(len(test_rows))
    
    train_rows.to_csv("Features_Clean_Normalized_Clinical_Data_time_LR_train.csv", index=False)
    test_rows.to_csv("Features_Clean_Normalized_Clinical_Data_time_LR_test.csv", index=False)
    
clinical_data_LR(Features,Clinical_Data)

#Passo1- ler ficheiro csv features + clinical data
#Passo2- eliminar colunas no clinical data: pacientes + 4tempos + DM+ OS
#Passo3- juntar 2 csv
#Passo4- dividir em teste + treino


def clinical_data_DM(Features,Clinical_Data):
        
    clean_data = Clinical_Data.drop(columns=['Patient','Time – diagnosis to CT sim (days)', 'Locoregional', 'Death'])
    

    
    data = Features.join(clean_data, how='inner')
    
    print(len(data))
    
    
    train_rows= data.loc[32:156]

    
    test_rows1 = data.loc[0:31]
    test_rows2 = data.loc[157:]
    
    test_rows3 = [test_rows1, test_rows2]
    
    
    test_rows = pd.concat(test_rows3)
    
    print(len(train_rows))
    print(len(test_rows1))
    print(len(test_rows2))
    print(len(test_rows))
    
    train_rows.to_csv("Features_Clean_Normalized_Clinical_Data_time_DM_train.csv", index=False)
    test_rows.to_csv("Features_Clean_Normalized_Clinical_Data_time_DM_test.csv", index=False)
    
clinical_data_DM(Features,Clinical_Data)

#Passo1- ler ficheiro csv features + clinical data
#Passo2- eliminar colunas no clinical data: pacientes + 4tempos + DM+ OS
#Passo3- juntar 2 csv
#Passo4- dividir em teste + treino

def clinical_data_OS(Features,Clinical_Data):
        
    clean_data = Clinical_Data.drop(columns=['Patient','Time – diagnosis to CT sim (days)'])
    
    data = Features.join(clean_data, how='inner')
       
    train_rows= data.loc[32:156]
   
    test_rows1 = data.loc[0:31]
    test_rows2 = data.loc[157:]
    
    test_rows3 = [test_rows1, test_rows2]
       
    test_rows = pd.concat(test_rows3)
    
    print(len(train_rows))
    print(len(test_rows1))
    print(len(test_rows2))
    print(len(test_rows))
    
    train_rows.to_csv("Features_Clean_Normalized_Clinical_Data_time_OS_train.csv", index=False)
    test_rows.to_csv("Features_Clean_Normalized_Clinical_Data_time_OS_test.csv", index=False)
    
clinical_data_OS(Features,Clinical_Data)
