==> Pre-Processing/Step5.2_Selection_FeatureImportance.ipynb <==
import pandas as pd

DM_test = pd.read_csv('Features_Clean_Normalized_DM_test.csv')
DM_train = pd.read_csv('Features_Clean_Normalized_DM_train.csv')
Feat_Imp= pd.read_csv('DM_XGBoost_gridSearch_FeatureImportance.csv')


def feat_imp_DM(Feat_Imp, DM_test, DM_train):

    
    columns_DM = list(DM_test.columns.values)
    
    important= Feat_Imp.loc[:1266]
    
    feat= important.loc[:,'feature']
    
    lista=[]
    for i in columns_DM:
        
        for j in feat:
            if i==j:
                lista.append(i)
                
         
    
    new_csv_test=DM_test.loc[:, DM_test.columns.isin(lista)]
    new_csv_train=DM_train.loc[:, DM_train.columns.isin(lista)]
    
    DM = DM_test['Distant']
    DM2= DM_train['Distant']
    
    DM_final_test = new_csv_test.join(DM, how='inner')
    DM_final_train = new_csv_train.join(DM2, how='inner')
    
    #print(DM_final_test)
    #print(DM_final_train)
                
    DM_final_test.to_csv('Features_Clean_Normalized_DM_test_Important_XGBoost.csv', index=False)
    DM_final_train.to_csv('Features_Clean_Normalized_DM_train_Important_XGBoost.csv', index=False)  
         


feat_imp_DM(Feat_Imp, DM_test, DM_train)
